+++ src/index.js
@@ -1,10 +1,1 @@
-let ngModuleName = 'formly';
-
-let angular = require('angular-fix');
-let ngModule = angular.module(ngModuleName, []);
-
-require('./providers')(ngModule);
-require('./services')(ngModule);
-require('./directives')(ngModule);
-
-module.exports = ngModuleName;
+module.exports = require('./index.common');