+++ src/providers/index.js
@@ -3,4 +3,5 @@ module.exports = ngModule => {
   require('./formlyConfig')(ngModule);
   require('./formlyVersion')(ngModule);
   require('./formlyErrorAndWarningsUrlPrefix')(ngModule);
+  require('./formlyValidationMessages')(ngModule);
 };