+++ src/run/index.js
@@ -1,0 +1,3 @@
+module.exports = ngModule => {
+  require('./formlyNgModelAttrsManipulator')(ngModule);
+};